Adding ServiceType model.Implementing CR of CRUD.

+++ ServicesOfBeautySalon/DAL/Advertisement.cs
@@ -14,10 +14,10 @@ namespace ServicesOfBeautySalon.DAL
     
     public partial class Advertisement
     {
-        public int ID { get; set; }
+        public long ID { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public int BeautySalonID { get; set; }
+        public long BeautySalonID { get; set; }
     
         public virtual BeautySalon BeautySalon { get; set; }
     }
